+++ packages/blobstorage/indexer/indexer.go
@@ -18,9 +18,9 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto/kzg4844"
 	"github.com/ethereum/go-ethereum/ethclient"
-	blobstorage "github.com/taikoxyz/taiko-mono/packages/blob-storage"
-	"github.com/taikoxyz/taiko-mono/packages/blob-storage/bindings/taikol1"
-	"github.com/taikoxyz/taiko-mono/packages/blob-storage/pkg/repo"
+	blobstorage "github.com/taikoxyz/taiko-mono/packages/blobstorage"
+	"github.com/taikoxyz/taiko-mono/packages/blobstorage/bindings/taikol1"
+	"github.com/taikoxyz/taiko-mono/packages/blobstorage/pkg/repo"
 	"github.com/urfave/cli/v2"
 	"golang.org/x/exp/slog"
 	"golang.org/x/sync/errgroup"
