+++ packages/blobstorage/scripts/abigen.sh
